+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## [Unreleased]
+
+### Changed
+
+- Testers (application and command) now automatically sets `decorated` to `False`
+
+
 ## [0.6.7] - 2018-06-25
 
 ### Fixed